Split ARCuboidView into ARCuboidEntity and ARParcelSceneView
