+++ WooCommerce/Classes/ViewRelated/Orders/Order Details/Shipping Labels/WooShipping Create Shipping Labels/Parcel Fitting Check/ARParcelSizingView.swift
@@ -36,7 +36,7 @@ struct ARParcelSizingView: View {
 
     var body: some View {
         ZStack {
-            ARCuboidView(
+            ARParcelSceneView(
                 dimensions: dimensionsInMeters(length: length, width: width, height: height),
                 isPlaced: $isPlaced,
                 resetTrigger: resetTrigger
